First push of Yale-styled CSS changes
This is the initial update to yale.css that adds in the yale.edu look-and-feel as well as the following:

- Swapping the right/left orientation of the search results and the search filters (filters now on left)
- Swapping the right/left orientation of the details and the collection organization (collection organization now on left)
- Elevation of breadcrumbs above title in record view

There are several "hacks" in this CSS that must be rectified in code to avoid sloppiness and potential breaks during upgrades.  Please see the comments inside the CSS file for details.