First crack at including YOP changes
Mark and Steve met with the Yale Office of the Printer to review the PUI interface.  They made many suggestions for aesthetic changes which we have begun to implement in this CSS file.

Major changes have been logged in Asana and commented in the file where appropriate.  The major areas touched in this commit:

- Collection/Sub-collection page alignments
- Button spacing
- "Suggest a Correction"
- Results display
- Results sort widgets
- Repository pages